CCEffects - Split some classes out of CCEffect_Private.h which was getting awfully big
