+++ cocos2d/CCEffect.m
@@ -43,284 +43,6 @@ static NSString* vertBase =
 @"	gl_Position = %@;\n"
 @"}\n";
 
-#pragma mark CCEffectFunction
-
-@implementation CCEffectFunction
-
--(id)initWithName:(NSString *)name body:(NSString*)body inputs:(NSArray*)inputs returnType:(NSString *)returnType
-{
-    if((self = [super init]))
-    {
-        NSAssert(inputs.count <= 1, @"Effect functions currently only support 0 or 1 inputs.");
-        
-        _body = [body copy];
-        _name = [name copy];
-        _inputs = [inputs copy];
-        _returnType = [returnType copy];
-
-        _inputString = @"void";
-        if (_inputs.count)
-        {
-            NSMutableString *tmpString = [[NSMutableString alloc] init];
-            for (CCEffectFunctionInput *input in _inputs)
-            {
-                [tmpString appendFormat:@"%@ %@", input.type, input.name];
-            }
-            _inputString = tmpString;
-        }
-        
-        return self;
-    }
-    
-    return self;
-}
-
-+(instancetype)functionWithName:(NSString*)name body:(NSString*)body inputs:(NSArray*)inputs returnType:(NSString*)returnType
-{
-    return [[self alloc] initWithName:name body:body inputs:inputs returnType:returnType];
-}
-
--(NSString*)function
-{
-    NSString* function = [NSString stringWithFormat:@"%@ %@(%@)\n{\n%@\n}", _returnType, _name, _inputString, _body];
-    return function;
-}
-
--(NSString*)callStringWithInputs:(NSArray*)inputs
-{
-    NSMutableString *callString = [[NSMutableString alloc] initWithFormat:@"%@(", _name];
-    for (NSString *input in inputs)
-    {
-        if ([inputs lastObject] != input)
-        {
-            [callString appendFormat:@"%@, ", input];
-        }
-        else
-        {
-            [callString appendFormat:@"%@", input];
-        }
-    }
-    [callString appendString:@")"];
-    
-    return callString;
-}
-
-@end
-
-#pragma mark CCEffectFunctionInput
-
-@implementation CCEffectFunctionInput
-
--(id)initWithType:(NSString*)type name:(NSString*)name initialSnippet:(NSString*)initialSnippet snippet:(NSString*)snippet
-{
-    if((self = [super init]))
-    {
-        _type = [type copy];
-        _name = [name copy];
-        _initialSnippet = [initialSnippet copy];
-        _snippet = [snippet copy];
-        return self;
-    }
-    
-    return self;
-}
-
-+(instancetype)inputWithType:(NSString*)type name:(NSString*)name initialSnippet:(NSString*)initialSnippet snippet:(NSString*)snippet
-{
-    return [[self alloc] initWithType:type name:name initialSnippet:initialSnippet snippet:snippet];
-}
-
-@end
-
-#pragma mark CCEffectUniform
-
-@implementation CCEffectUniform
-
--(id)initWithType:(NSString*)type name:(NSString*)name value:(NSValue*)value
-{
-    if((self = [super init]))
-    {
-        _name = [name copy];
-        _type = [type copy];
-        _value = value;
-        
-        return self;
-    }
-    
-    return self;
-}
-
-+(instancetype)uniform:(NSString*)type name:(NSString*)name value:(NSValue*)value
-{
-    return [[self alloc] initWithType:type name:name value:value];
-}
-
--(NSString*)declaration
-{
-    NSString* declaration = [NSString stringWithFormat:@"uniform %@ %@;", _type, _name];
-    return declaration;
-}
-
-@end
-
-#pragma mark CCEffectVarying
-
-@implementation CCEffectVarying
-
--(id)initWithType:(NSString*)type name:(NSString*)name
-{
-    if((self = [self initWithType:type name:name count:0]))
-    {
-        return self;
-    }
-    
-    return self;
-}
-
-+(instancetype)varying:(NSString*)type name:(NSString*)name
-{
-    return [[self alloc] initWithType:type name:name];
-}
-
--(id)initWithType:(NSString*)type name:(NSString*)name count:(NSInteger)count
-{
-    if((self = [super init]))
-    {
-        _name = name;
-        _type = type;
-        _count = count;
-        
-        return self;
-    }
-    
-    return self;
-}
-
-+(instancetype)varying:(NSString*)type name:(NSString*)name count:(NSInteger)count
-{
-    return [[self alloc] initWithType:type name:name count:count];
-}
-
-
--(NSString*)declaration
-{
-    NSString* declaration;
-
-    if(_count == 0)
-        declaration = [NSString stringWithFormat:@"varying %@ %@;", _type, _name];
-    else
-        declaration = [NSString stringWithFormat:@"varying %@ %@[%lu];", _type, _name, (long)_count];
-    
-    return declaration;
-}
-
-@end
-
-
-#pragma mark CCEffectRenderPassInputs
-
-@implementation CCEffectRenderPassInputs
-
--(id)init
-{
-    return [super init];
-}
-
-@end
-
-
-#pragma mark CCEffectRenderPass
-
-@implementation CCEffectRenderPass
-
--(id)init
-{
-    return [self initWithIndex:0];
-}
-
--(id)initWithIndex:(NSUInteger)indexInEffect
-{
-    if((self = [super init]))
-    {
-        _indexInEffect = indexInEffect;
-        
-        _texCoord1Mapping = CCEffectTexCoordMapPreviousPassTex;
-        _texCoord2Mapping = CCEffectTexCoordMapCustomTex;
-        
-        _beginBlocks = @[[^(CCEffectRenderPass *pass, CCEffectRenderPassInputs *passInputs){} copy]];
-        _endBlocks = @[[^(CCEffectRenderPass *pass, CCEffectRenderPassInputs *passInputs){} copy]];
-
-        CCEffectRenderPassUpdateBlock updateBlock = ^(CCEffectRenderPass *pass, CCEffectRenderPassInputs *passInputs){
-            if (passInputs.needsClear)
-            {
-                [passInputs.renderer enqueueClear:GL_COLOR_BUFFER_BIT color:[CCColor clearColor].glkVector4 depth:0.0f stencil:0 globalSortOrder:NSIntegerMin];
-            }
-            [pass enqueueTriangles:passInputs];
-        };
-        _updateBlocks = @[[updateBlock copy]];
-        _blendMode = [CCBlendMode premultipliedAlphaMode];
-        
-        return self;
-    }
-    
-    return self;
-}
-
--(instancetype)copyWithZone:(NSZone *)zone
-{
-	CCEffectRenderPass *newPass = [[CCEffectRenderPass allocWithZone:zone] initWithIndex:_indexInEffect];
-    newPass.texCoord1Mapping = _texCoord1Mapping;
-    newPass.texCoord2Mapping = _texCoord2Mapping;
-    newPass.blendMode = _blendMode;
-    newPass.shader = _shader;
-    newPass.beginBlocks = _beginBlocks;
-    newPass.updateBlocks = _updateBlocks;
-    newPass.endBlocks = _endBlocks;
-    newPass.debugLabel = _debugLabel;
-    return newPass;
-}
-
--(void)begin:(CCEffectRenderPassInputs *)passInputs
-{
-    for (CCEffectRenderPassBeginBlock block in _beginBlocks)
-    {
-        block(self, passInputs);
-    }
-}
-
--(void)update:(CCEffectRenderPassInputs *)passInputs
-{
-    for (CCEffectRenderPassUpdateBlock block in _updateBlocks)
-    {
-        block(self, passInputs);
-    }
-}
-
--(void)end:(CCEffectRenderPassInputs *)passInputs
-{
-    for (CCEffectRenderPassUpdateBlock block in _endBlocks)
-    {
-        block(self, passInputs);
-    }
-}
-
--(void)enqueueTriangles:(CCEffectRenderPassInputs *)passInputs
-{
-    CCRenderState *renderState = [CCRenderState renderStateWithBlendMode:_blendMode shader:_shader shaderUniforms:passInputs.shaderUniforms copyUniforms:YES];
-    
-    GLKMatrix4 transform = passInputs.transform;
-    CCRenderBuffer buffer = [passInputs.renderer enqueueTriangles:2 andVertexes:4 withState:renderState globalSortOrder:0];
-
-    CCRenderBufferSetVertex(buffer, 0, CCVertexApplyTransform(passInputs.verts.bl, &transform));
-	CCRenderBufferSetVertex(buffer, 1, CCVertexApplyTransform(passInputs.verts.br, &transform));
-	CCRenderBufferSetVertex(buffer, 2, CCVertexApplyTransform(passInputs.verts.tr, &transform));
-	CCRenderBufferSetVertex(buffer, 3, CCVertexApplyTransform(passInputs.verts.tl, &transform));
-	
-	CCRenderBufferSetTriangle(buffer, 0, 0, 1, 2);
-	CCRenderBufferSetTriangle(buffer, 1, 0, 2, 3);
-}
-
-@end
 
 #pragma mark CCEffectImpl
