+++ cocos2d/CCEffectFunction.h
@@ -1,0 +1,38 @@
+//
+//  CCEffectFunction.h
+//  cocos2d
+//
+//  Created by Thayer J Andrews on 3/5/15.
+//
+//
+
+#import "CCEffect.h"
+
+@interface CCEffectFunction : NSObject
+
+@property (nonatomic, readonly) NSString* body;
+@property (nonatomic, readonly) NSString* name;
+@property (nonatomic, readonly) NSArray* inputs;
+@property (nonatomic, readonly) NSString* inputString;
+@property (nonatomic, readonly) NSString* returnType;
+@property (nonatomic, readonly) NSString* function;
+
+-(id)initWithName:(NSString*)name body:(NSString*)body inputs:(NSArray*)inputs returnType:(NSString*)returnType;
++(instancetype)functionWithName:(NSString*)name body:(NSString*)body inputs:(NSArray*)inputs returnType:(NSString*)returnType;
+
+-(NSString*)callStringWithInputs:(NSArray*)inputs;
+
+@end
+
+@interface CCEffectFunctionInput : NSObject
+
+@property (nonatomic, readonly) NSString* type;
+@property (nonatomic, readonly) NSString* name;
+@property (nonatomic, readonly) NSString* initialSnippet;
+@property (nonatomic, readonly) NSString* snippet;
+
+-(id)initWithType:(NSString*)type name:(NSString*)name initialSnippet:(NSString*)initialSnippet snippet:(NSString*)snippet;
++(instancetype)inputWithType:(NSString*)type name:(NSString*)name initialSnippet:(NSString*)initialSnippet snippet:(NSString*)snippet;
+
+@end
+
