+++ cocos2d/CCEffectFunction.m
@@ -1,0 +1,99 @@
+//
+//  CCEffectFunction.m
+//  cocos2d
+//
+//  Created by Thayer J Andrews on 3/5/15.
+//
+//
+
+#import "CCEffectFunction.h"
+
+#pragma mark CCEffectFunction
+
+@implementation CCEffectFunction
+
+-(id)initWithName:(NSString *)name body:(NSString*)body inputs:(NSArray*)inputs returnType:(NSString *)returnType
+{
+    if((self = [super init]))
+    {
+        NSAssert(inputs.count <= 1, @"Effect functions currently only support 0 or 1 inputs.");
+        
+        _body = [body copy];
+        _name = [name copy];
+        _inputs = [inputs copy];
+        _returnType = [returnType copy];
+        
+        _inputString = @"void";
+        if (_inputs.count)
+        {
+            NSMutableString *tmpString = [[NSMutableString alloc] init];
+            for (CCEffectFunctionInput *input in _inputs)
+            {
+                [tmpString appendFormat:@"%@ %@", input.type, input.name];
+            }
+            _inputString = tmpString;
+        }
+        
+        return self;
+    }
+    
+    return self;
+}
+
++(instancetype)functionWithName:(NSString*)name body:(NSString*)body inputs:(NSArray*)inputs returnType:(NSString*)returnType
+{
+    return [[self alloc] initWithName:name body:body inputs:inputs returnType:returnType];
+}
+
+-(NSString*)function
+{
+    NSString* function = [NSString stringWithFormat:@"%@ %@(%@)\n{\n%@\n}", _returnType, _name, _inputString, _body];
+    return function;
+}
+
+-(NSString*)callStringWithInputs:(NSArray*)inputs
+{
+    NSMutableString *callString = [[NSMutableString alloc] initWithFormat:@"%@(", _name];
+    for (NSString *input in inputs)
+    {
+        if ([inputs lastObject] != input)
+        {
+            [callString appendFormat:@"%@, ", input];
+        }
+        else
+        {
+            [callString appendFormat:@"%@", input];
+        }
+    }
+    [callString appendString:@")"];
+    
+    return callString;
+}
+
+@end
+
+#pragma mark CCEffectFunctionInput
+
+@implementation CCEffectFunctionInput
+
+-(id)initWithType:(NSString*)type name:(NSString*)name initialSnippet:(NSString*)initialSnippet snippet:(NSString*)snippet
+{
+    if((self = [super init]))
+    {
+        _type = [type copy];
+        _name = [name copy];
+        _initialSnippet = [initialSnippet copy];
+        _snippet = [snippet copy];
+        return self;
+    }
+    
+    return self;
+}
+
++(instancetype)inputWithType:(NSString*)type name:(NSString*)name initialSnippet:(NSString*)initialSnippet snippet:(NSString*)snippet
+{
+    return [[self alloc] initWithType:type name:name initialSnippet:initialSnippet snippet:snippet];
+}
+
+@end
+
