+++ cocos2d/CCEffectUniform.m
@@ -1,0 +1,41 @@
+//
+//  CCEffectUniform.m
+//  cocos2d
+//
+//  Created by Thayer J Andrews on 3/5/15.
+//
+//
+
+#import "CCEffectUniform.h"
+
+#pragma mark CCEffectUniform
+
+@implementation CCEffectUniform
+
+-(id)initWithType:(NSString*)type name:(NSString*)name value:(NSValue*)value
+{
+    if((self = [super init]))
+    {
+        _name = [name copy];
+        _type = [type copy];
+        _value = value;
+        
+        return self;
+    }
+    
+    return self;
+}
+
++(instancetype)uniform:(NSString*)type name:(NSString*)name value:(NSValue*)value
+{
+    return [[self alloc] initWithType:type name:name value:value];
+}
+
+-(NSString*)declaration
+{
+    NSString* declaration = [NSString stringWithFormat:@"uniform %@ %@;", _type, _name];
+    return declaration;
+}
+
+@end
+
