+++ cocos2d/CCEffect_Private.h
@@ -7,7 +7,11 @@
 //
 
 #import "CCEffect.h"
+#import "CCEffectFunction.h"
+#import "CCEffectRenderPass.h"
 #import "CCEffectStackProtocol.h"
+#import "CCEffectUniform.h"
+#import "CCEffectVarying.h"
 
 
 extern NSString * const CCShaderUniformPreviousPassTexture;
@@ -47,118 +51,6 @@ typedef NS_ENUM(NSUInteger, CCEffectFunctionStitchFlags)
     CCEffectFunctionStitchAfter      = 1 << 1,
     CCEffectFunctionStitchBoth       = (CCEffectFunctionStitchBefore | CCEffectFunctionStitchAfter),
 };
-
-typedef NS_ENUM(NSUInteger, CCEffectTexCoordMapping)
-{
-    CCEffectTexCoordMapMainTex              = 0,
-    CCEffectTexCoordMapPreviousPassTex      = 1,
-    CCEffectTexCoordMapCustomTex            = 2,
-    CCEffectTexCoordMapCustomTexNoTransform = 3
-};
-
-@interface CCEffectFunction : NSObject
-
-@property (nonatomic, readonly) NSString* body;
-@property (nonatomic, readonly) NSString* name;
-@property (nonatomic, readonly) NSArray* inputs;
-@property (nonatomic, readonly) NSString* inputString;
-@property (nonatomic, readonly) NSString* returnType;
-@property (nonatomic, readonly) NSString* function;
-
--(id)initWithName:(NSString*)name body:(NSString*)body inputs:(NSArray*)inputs returnType:(NSString*)returnType;
-+(instancetype)functionWithName:(NSString*)name body:(NSString*)body inputs:(NSArray*)inputs returnType:(NSString*)returnType;
-
--(NSString*)callStringWithInputs:(NSArray*)inputs;
-
-@end
-
-@interface CCEffectFunctionInput : NSObject
-
-@property (nonatomic, readonly) NSString* type;
-@property (nonatomic, readonly) NSString* name;
-@property (nonatomic, readonly) NSString* initialSnippet;
-@property (nonatomic, readonly) NSString* snippet;
-
--(id)initWithType:(NSString*)type name:(NSString*)name initialSnippet:(NSString*)initialSnippet snippet:(NSString*)snippet;
-+(instancetype)inputWithType:(NSString*)type name:(NSString*)name initialSnippet:(NSString*)initialSnippet snippet:(NSString*)snippet;
-
-@end
-
-@interface CCEffectUniform : NSObject
-
-@property (nonatomic, readonly) NSString* name;
-@property (nonatomic, readonly) NSString* type;
-@property (nonatomic, readonly) NSString* declaration;
-@property (nonatomic, readonly) NSValue* value;
-
--(id)initWithType:(NSString*)type name:(NSString*)name value:(NSValue*)value;
-+(instancetype)uniform:(NSString*)type name:(NSString*)name value:(NSValue*)value;
-
-@end
-
-@interface CCEffectVarying : NSObject
-
-@property (nonatomic, readonly) NSString* name;
-@property (nonatomic, readonly) NSString* type;
-@property (nonatomic, readonly) NSString* declaration;
-@property (nonatomic, readonly) NSInteger count;
-
--(id)initWithType:(NSString*)type name:(NSString*)name;
--(id)initWithType:(NSString*)type name:(NSString*)name count:(NSInteger)count;
-+(instancetype)varying:(NSString*)type name:(NSString*)name;
-+(instancetype)varying:(NSString*)type name:(NSString*)name count:(NSInteger)count;
-
-@end
-
-
-
-@interface CCEffectRenderPassInputs : NSObject
-
-@property (nonatomic, assign) NSInteger renderPassId;
-@property (nonatomic, strong) CCRenderer* renderer;
-@property (nonatomic, strong) CCSprite *sprite;
-@property (nonatomic, assign) CCSpriteVertexes verts;
-@property (nonatomic, strong) CCTexture *previousPassTexture;
-@property (nonatomic, assign) GLKMatrix4 transform;
-@property (nonatomic, assign) GLKMatrix4 ndcToNodeLocal;
-@property (nonatomic, assign) GLKVector2 texCoord1Center;
-@property (nonatomic, assign) GLKVector2 texCoord1Extents;
-@property (nonatomic, assign) GLKVector2 texCoord2Center;
-@property (nonatomic, assign) GLKVector2 texCoord2Extents;
-@property (nonatomic, strong) NSMutableDictionary* shaderUniforms;
-@property (nonatomic, assign) BOOL needsClear;
-
-@end
-
-
-@class CCEffectRenderPass;
-
-typedef void (^CCEffectRenderPassBeginBlock)(CCEffectRenderPass *pass, CCEffectRenderPassInputs *passInputs);
-typedef void (^CCEffectRenderPassUpdateBlock)(CCEffectRenderPass *pass, CCEffectRenderPassInputs *passInputs);
-typedef void (^CCEffectRenderPassEndBlock)(CCEffectRenderPass *pass, CCEffectRenderPassInputs *passInputs);
-
-@interface CCEffectRenderPass : NSObject
-
-@property (nonatomic, readonly) NSUInteger indexInEffect;
-@property (nonatomic, assign) CCEffectTexCoordMapping texCoord1Mapping;
-@property (nonatomic, assign) CCEffectTexCoordMapping texCoord2Mapping;
-@property (nonatomic, strong) CCBlendMode* blendMode;
-@property (nonatomic, strong) CCShader* shader;
-@property (nonatomic, strong) NSDictionary* uniformTranslationTable;
-@property (nonatomic, copy) NSArray* beginBlocks;
-@property (nonatomic, copy) NSArray* updateBlocks;
-@property (nonatomic, copy) NSArray* endBlocks;
-@property (nonatomic, copy) NSString *debugLabel;
-
--(id)initWithIndex:(NSUInteger)indexInEffect;
-
--(void)begin:(CCEffectRenderPassInputs *)passInputs;
--(void)update:(CCEffectRenderPassInputs *)passInputs;
--(void)end:(CCEffectRenderPassInputs *)passInputs;
--(void)enqueueTriangles:(CCEffectRenderPassInputs *)passInputs;
-
-@end
-
 
 @class CCEffectImpl;
